Adding reaction for hw home button
When the camera activity is shown, pushing hw home button causes the application to minimize. When application is restored the camera activity is not shown but scanner.launched is true. Trying to reset or stop scanner or even dispose and reinitialise then launch fails.
I guess that adding the same reaction as for hw back should solve the problem.